+++ packages/did-ui-react/src/components/SocialLogin/index.tsx
@@ -17,7 +17,7 @@ import AccountRecommendGroup from '../AccountRecommendGroup';
 import TermsOfServiceItem from '../TermsOfServiceItem';
 import { CreateWalletType, LoginFinishWithoutPin, Theme } from '../types';
 import useSocialLogin from '../../hooks/useSocialLogin';
-import { errorTip, handleErrorMessage, setLoading } from '../../utils';
+import { did, errorTip, handleErrorMessage, setLoading } from '../../utils';
 import './index.less';
 import { TotalAccountsInfo } from '../../constants/socialLogin';
 import { AccountLoginList, SocialLoginList, Web2LoginList } from '../../constants/guardian';
@@ -25,6 +25,7 @@ import { AccountType } from '@portkey/services';
 import { useComputeIconCountPreRow } from '../../hooks/login';
 import UpgradedPortkeyTip from '../UpgradedPortkeyTip';
 import { getAccessTokenInDappTelegram, getTelegramStartParam } from '../../utils/telegram';
+import { PORTKEY_LOGIN_STORAGE_KEY } from '../../constants/storage';
 
 interface SocialLoginProps {
   type: RegisterType;
@@ -116,18 +117,20 @@ export default function SocialLogin({
   const handleLoginAfterAuthWithInTelegram = useCallback(async () => {
     // TODO tg test
     const { startParam } = getTelegramStartParam();
-    // const startParam = '1712049223890_986012';
+    // const startParam = '1712052456778_38986';
     console.log('====startParam', startParam);
     if (startParam) {
       clearInterval(timerRef.current);
       const decodeResult = await getAccessTokenInDappTelegram(startParam);
       console.log('===res.data', decodeResult);
 
+      await did.config.storageMethod.removeItem(`${PORTKEY_LOGIN_STORAGE_KEY}_Telegram`);
+
       // get AccessToken from socket
       const result = {
         type: 'Telegram' as ISocialLogin,
         data: {
-          accessToken: decodeResult?.token || '',
+          accessToken: decodeResult || '',
         },
       };
       console.log('=== result', result);