+++ packages/did-ui-react/src/components/TelegramLoginButton/index.tsx
@@ -57,6 +57,7 @@ export default function TelegramLoginButton({ className, onBeforeBack }: Telegra
   });
 
   const handleTGAuth = useCallback(async () => {
+    // TODO tg test
     if (!telegramInitData.current?.user || !telegramInitData.current?.auth_date || !telegramInitData.current?.hash) {
       return singleMessage.info('Please wait');
     }
@@ -65,6 +66,7 @@ export default function TelegramLoginButton({ className, onBeforeBack }: Telegra
       setLoading(true);
 
       const { startParam } = getTelegramStartParam();
+      // const startParam = '1712049223890_986012';
       if (!startParam) throw Error('No loginId');
 
       await getAccessTokenAndOpenPortkeyWebapp({
@@ -79,10 +81,6 @@ export default function TelegramLoginButton({ className, onBeforeBack }: Telegra
     } finally {
       setLoading(false);
     }
-
-    // did.services.getRegisterInfo({
-    //   loginGuardianIdentifier: userId,
-    // });
   }, [onBeforeBack]);
 
   return (